requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
distro==1.9.0
ecdsa==0.19.1
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.99.9
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-http-client==3.3.7
sendgrid==6.12.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
Werkzeug==3.1.3
WTForms==3.2.1
